fix: Feign 기능 수정
- Member 조회 로직 공통 사용을 위해 Dictionary 형식으로 변환하여 응답
- 이미지 삭제시 Member API에
AWS이미지 삭제 호출을 요청

+++ community-service/src/main/java/com/freediving/communityservice/application/service/ArticleService.java
@@ -18,8 +18,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 import org.springframework.util.ObjectUtils;
 
-import com.freediving.common.response.ResponseJsonObject;
-import com.freediving.common.response.dto.member.MemberFindUserResponse;
+import com.freediving.common.handler.exception.BuddyMeException;
 import com.freediving.common.response.enumerate.ServiceStatusCode;
 import com.freediving.communityservice.adapter.out.dto.article.ArticleBriefDto;
 import com.freediving.communityservice.adapter.out.dto.article.ArticleContent;
@@ -170,29 +169,12 @@ public class ArticleService implements ArticleUseCase {
 			.distinct()
 			.toList();
 
-		ResponseJsonObject<List<MemberFindUserResponse>> memberInfoList = memberFeignPort.findUserListByUserIds(
-			articleOwnerIds, true);
+		Map<Long, UserInfo> userMap = memberFeignPort.getUserMapByUserIds(articleOwnerIds, true);
 
-		if (ServiceStatusCode.OK.getCode() == memberInfoList.getCode()) {
-			Map<Long, MemberFindUserResponse> userDict = memberInfoList.getData().stream()
-				.collect(Collectors.toUnmodifiableMap(
-					MemberFindUserResponse::getUserId, m -> m
-				));
-
-			for (ArticleBriefDto articleBriefDto : pagingArticleList.getContent()) {
-				if (userDict.containsKey(articleBriefDto.getCreatedBy())) {
-					MemberFindUserResponse user = userDict.get(articleBriefDto.getCreatedBy());
-
-					articleBriefDto.setUserInfo(
-						UserInfo.builder()
-							.nickname(user.getNickname())
-							.profileImgUrl(user.getProfileImgUrl())
-							.licenseInfo(user.getLicenseInfo())
-							.build()
-					);
-				}
-			}
+		for (ArticleBriefDto articleBriefDto : pagingArticleList.getContent()) {
+			articleBriefDto.setUserInfo(userMap.get(articleBriefDto.getCreatedBy()));
 		}
+
 		return pagingArticleList;
 
 	}
@@ -292,8 +274,12 @@ public class ArticleService implements ArticleUseCase {
 			}
 		}
 		imageWritePort.saveImages(article.getId(), article.getCreatedBy(), article.getCreatedAt(), imagesToCreate);
-		//TODO: Member에 AWS 삭제 요청
-		imageDeletePort.deleteAllByArticleIdAndUrlIn(article.getId(), imageUrlsToDelete);
+
+		if (!imageUrlsToDelete.isEmpty()) {
+			memberFeignPort.deleteImageByUrls(imageUrlsToDelete);
+			imageDeletePort.deleteAllByArticleIdAndUrlIn(article.getId(), imageUrlsToDelete);
+		}
+
 		imageEditPort.editAllImageSortNumber(article.getId(), imageMapToUpdate);
 	}
 
@@ -304,15 +290,17 @@ public class ArticleService implements ArticleUseCase {
 
 		Article article = articleReadPort.readArticle(command.getBoardType(), command.getArticleId(), isShowAll);
 		if (Objects.isNull(article)) {
-			throw new IllegalArgumentException("해당하는 게시글이 없습니다.");
+			throw new BuddyMeException(ServiceStatusCode.BAD_REQUEST, "해당하는 게시글이 없습니다.");
 		}
 		article.checkHasOwnership(command.getUserProvider().getRequestUserId());
 
+		List<ImageResponse> articleImages = imageReadPort.getImageListByArticle(article.getId());
+
 		articleDeletePort.markDeleted(command);
 		commentDeletePort.markDeletedByArticleId(article.getId());
-		imageDeletePort.deleteAllByArticleId(article.getId());
-		//TODO: Member로 AWS Image삭제 요청
 
+		memberFeignPort.deleteImageByUrls(articleImages.stream().map(ImageResponse::url).toList());
+		imageDeletePort.deleteAllByArticleId(article.getId());
 		return article.getId();
 	}
